Modularizing commit
Novos módulos aplicados tendo em vista os mesmos comportamentos anteriores
