+++ templates/base.html
@@ -19,7 +19,7 @@
 	  <header>
 		<nav class="navbar navbar-expand-lg orange">
 		  <div class="container-fluid">
-			<a class="navbar-brand" href="{{url_for('index')}}">Cãopanheiro</a>
+			<a class="navbar-brand" href="{{url_for('bp.index')}}">Cãopanheiro</a>
 			<button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNavAltMarkup" aria-controls="navbarNavAltMarkup" aria-expanded="false" aria-label="Toggle navigation">
 			  <span class="navbar-toggler-icon"></span>
 			</button>
@@ -43,8 +43,8 @@
 					 <li><a class="dropdown-item" href="#">Informaçoes pessoais</a></li>
 <!--					 altera preferencias e cadastro de acordo com o perfil de pessoa-->
 <!--					 adicionar uma pagina de preferencias existentes com botao de alteração-->
-					 <li><a class="dropdown-item" href="{{url_for('preferences')}}">Preferencias</a></li>
-					 <li><a class="dropdown-item" href="{{url_for('dogs')}}">Cadastro de cachorros</a></li>
+					 <li><a class="dropdown-item" href="{{url_for('bp.preferences')}}">Preferencias</a></li>
+					 <li><a class="dropdown-item" href="{{url_for('bp.dogs')}}">Cadastro de cachorros</a></li>
 					 <li><hr class="dropdown-divider"></li>
 					 <li><a class="dropdown-item" href="#">Sair</a></li>
 				 </ul>
