+++ templates/login.html
@@ -23,7 +23,7 @@
         </div>
     </form>
 
-    <a href="{{ url_for('user_register') }}">Cadastre-se</a>
+    <a href="{{ url_for('bp.user_register') }}">Cadastre-se</a>
 
 </body>
 </html>
