+++ templates/new_dog.html
@@ -1,7 +1,7 @@
 {% extends 'base.html' %}
 {% block conteudo %}
     <section>
-        <form action="{{url_for('add_dog')}}" method="POST" class="form-control">
+        <form action="{{url_for('bp.add_dog')}}" method="POST" class="form-control">
             <h1 class="my-3" style="text-align: center;">Cadastro do Cachorro</h1>
             <div class="mb-3 form-group">
                 <label for="name" class="form-label">Nome</label>
