+++ templates/preferences.html
@@ -1,7 +1,7 @@
 {% extends 'base.html' %}
 {% block conteudo %}
     <section>
-        <form action="{{url_for('preferences')}}" method="POST" class="form-control">
+        <form action="{{url_for('bp.preferences')}}" method="POST" class="form-control">
             <h1 class="my-3" style="text-align: center;">Questionário de preferências para adoção responsável</h1>
 
             <div class="mb-3 form-group">
